Vectorize separation() to support array inputs
The separation() function now handles both scalar and array RA/Dec inputs,
enabling illumination_and_power() to work with arrays of times.

Changes:
- radec2vec() returns (n, 3) for array input, (3,) for scalar
- separation() uses vectorized path for arrays, original path for scalars
- Add float() casts in callers that expect scalar returns
- Enable test_illumination_and_power_with_array_times (was skipped)

+++ conops/simulation/passes.py
@@ -144,9 +144,11 @@ class Pass(BaseModel):
         # Convert to radians
 
         return (
-            separation(
-                [spacecraft_ra * DTOR, spacecraft_dec * DTOR],
-                [target_ra * DTOR, target_dec * DTOR],
+            float(
+                separation(
+                    [spacecraft_ra * DTOR, spacecraft_dec * DTOR],
+                    [target_ra * DTOR, target_dec * DTOR],
+                )
             )
             / DTOR
         )
